Even push en commit voor pauze, verder gewerkt aan update playlist

+++ src/main/java/nl/han/dea/marc/datasources/TracksDAO.java
@@ -27,7 +27,7 @@ public class TracksDAO {
                     rs.getInt(1),
                     rs.getString(2),
                     rs.getString(3),
-                    rs.getDouble(4),
+                    rs.getInt(4),
                     rs.getString(5),
                     rs.getInt(6),
                     rs.getString(7),
